+++ tests/src/rules/default.js
@@ -221,6 +221,17 @@ context('TypeScript', function () {
           },
         }),
         test({
+          code: `import Foo from "./typescript-export-react-test-renderer"`,
+          parser: parser,
+          settings: {
+            'import/parsers': { [parser]: ['.ts'] },
+            'import/resolver': { 'eslint-import-resolver-typescript': true },
+          },
+          parserOptions: {
+            tsconfigRootDir: path.resolve(__dirname, '../../files/typescript-export-react-test-renderer/'),
+          },
+        }),
+        test({
           code: `import foobar from "./typescript-export-assign-property"`,
           parser: parser,
           settings: {